requirements.txt
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
astroid==3.3.8
asttokens==3.0.0
cachetools==5.5.0
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
detect-secrets==1.5.0
dill==0.3.9
distlib==0.3.9
distro==1.9.0
executing==2.1.0
filelock==3.16.1
fonttools==4.55.3
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
identify==2.6.4
idna==3.10
importTime==1.0.3.2
ipykernel==6.29.5
ipython==8.31.0
isort==5.13.2
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.2.1
oauthlib==3.2.2
openai==1.58.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pipdeptree==2.24.0
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pylint==3.3.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
svgwrite==1.4.3
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Tree==0.2.4
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.28.0
wcwidth==0.2.13
